requirements.txt
Django>=2.0.5
django-filters>=0.2.1
djangorestframework>=3.8.2
djangorestframework-jwt>=1.11.0
drfaddons>=0.0.2
PyJWT>=1.6.4
pytz>=2018.4
